+++ src/main/java/Main.java
@@ -1,12 +1,12 @@
-import Entities.DataPoint;
-import Entities.Metric;
+import entities.DataPoint;
+import entities.Metric;
 
-import Entities.MetricStorage;
-import Models.ImportRequestModel;
-import Presenters.DataImportPresenter;
-import Presenters.DataImportPresenterOutputBoundary;
-import UseCases.DataImportInputBoundary;
-import UseCases.DataImporter;
+import entities.MetricStorage;
+import models.ImportRequestModel;
+import presenters.DataImportPresenter;
+import presenters.DataImportPresenterOutputBoundary;
+import use_cases.DataImportInputBoundary;
+import use_cases.DataImporter;
 
 import java.io.IOException;
