+++ src/main/java/entities/DataPoint.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
