+++ src/main/java/entities/MetricStorageInterface.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 import java.io.File;
 import java.util.ArrayList;
