+++ src/main/java/presenters/DataExportPresenterOutputBoundary.java
@@ -1,6 +1,6 @@
-package Presenters;
+package presenters;
 
-import Models.ExportResponseModel;
+import models.ExportResponseModel;
 
 public interface DataExportPresenterOutputBoundary {
     ExportResponseModel prepareSuccessView();
